WIP: outmacho: Set a default of macOS 10.4.0 or macOS 10.5.0 if nothing is set
In Xcode 15.0, the new linker warns, if no min OS has been set:

    ld: warning: no platform load command found in 'obj.o', assuming: macOS

A similar warning appeared in Xcode 12.0, with warnings like this:

    Building for macOS, but linking in object file (obj.o) built for free standing

In that case, the warning seemed to have been silenced in some
later update.

By defaulting to macOS, the warning goes away in most cases;
however if building for e.g. the iOS Simulator, this instead
causes a hard error:

    ld: building for 'iOS-simulator', but linking in object file (obj.o) built for 'macOS'

For such cases, this would force users to pass the new option like
"--macho-min-os ios17.0-simulator".

Signed-off-by: Martin Storsjö <martin@martin.st>